requirements.txt
docx2pdf==0.1.8
numpy==1.23.5
pandas==2.2.2
tensorflow==2.12.0
tensorflow-docs @ git+https://github.com/tensorflow/docs.git@8246c1cba26bb07d9c02c165d83936c5b50825ca
scikit-surprise==1.1.4
scikit-learn==1.1.3
requests==2.32.3
Faker==25.3.0
pydantic==1.10.9
python-dotenv==1.0.1
mysql-connector-python==9.0.0